pyproject.toml
[project]
name = "trainable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.3",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "flask-ckeditor>=1.0.0",
    "bleach>=6.1.0",
    "ruff>=0.6.9",
]
